+++ CHANGELOG.md
@@ -11,6 +11,12 @@ Changelog
 
 _Note: Gaps between patch versions are faulty, broken or test releases._
 
+## v3.??.?? (2022-??-??)
+
+#### :bug: Bug Fix
+
+* Fixed setting of new deep properties of the `accessors` engine of the `watch` module
+
 ## v3.87.0 (2022-08-08)
 
 #### :rocket: New Feature