+++ CHANGELOG.md
@@ -15,7 +15,7 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 
 #### :bug: Bug Fix
 
-* Fixed deep setting of new properties of the `accessors` engine of the `watch` module
+* Fixed setting new properties of a watchable object by a deep path using the accessor-based engine `core/object/watch`
 
 ## v3.87.0 (2022-08-08)